+++ examples/data-sources/nxos_user/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_user" "example" {
-  name = "user1"
-}