+++ examples/data-sources/nxos_user_management/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_user_management" "example" {
+}