+++ examples/resources/nxos_user/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_user.example
-  identity = {
-    "name" : "<name>"
-  }
-}