+++ examples/resources/nxos_user/resource.tf
@@ -1,9 +1,0 @@
-resource "nxos_user" "example" {
-  name                     = "user1"
-  allow_expired            = "yes"
-  password                 = "password123"
-  password_encryption_type = "clear"
-  roles = [{
-    name = "network-operator"
-  }]
-}