+++ examples/resources/nxos_user_management/import-by-identity.tf
@@ -1,0 +1,5 @@
+import {
+  to = nxos_user_management.example
+  identity = {
+  }
+}