+++ examples/resources/nxos_user_management/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = nxos_user_management.example
+  id = ""
+}