+++ internal/provider/data_source_nxos_user_management.go
@@ -38,26 +38,26 @@ import (
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
-	_ datasource.DataSource              = &UserDataSource{}
-	_ datasource.DataSourceWithConfigure = &UserDataSource{}
+	_ datasource.DataSource              = &UserManagementDataSource{}
+	_ datasource.DataSourceWithConfigure = &UserManagementDataSource{}
 )
 
-func NewUserDataSource() datasource.DataSource {
-	return &UserDataSource{}
+func NewUserManagementDataSource() datasource.DataSource {
+	return &UserManagementDataSource{}
 }
 
-type UserDataSource struct {
+type UserManagementDataSource struct {
 	data *NxosProviderData
 }
 
-func (d *UserDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_user"
+func (d *UserManagementDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_user_management"
 }
 
-func (d *UserDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+func (d *UserManagementDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewResourceDescription("This data source can read the User configuration.", "aaaUser", "Security%20and%20Policing/aaa:User/").AddAdditionalDocs([]string{"aaaUserDomain", "aaaUserRole"}, []string{"Security%20and%20Policing/aaa:UserDomain/", "Security%20and%20Policing/aaa:UserRole/"}).String,
+		MarkdownDescription: helpers.NewResourceDescription("This data source can read the user management configuration.", "aaaUserEp", "Security%20and%20Policing/aaa:UserEp/").AddAdditionalDocs([]string{"aaaUser", "aaaUserDomain", "aaaUserRole"}, []string{"Security%20and%20Policing/aaa:User/", "Security%20and%20Policing/aaa:UserDomain/", "Security%20and%20Policing/aaa:UserRole/"}).String,
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{
@@ -68,30 +68,38 @@ func (d *UserDataSource) Schema(ctx context.Context, req datasource.SchemaReques
 				MarkdownDescription: "The distinguished name of the object.",
 				Computed:            true,
 			},
-			"name": schema.StringAttribute{
-				MarkdownDescription: "User name.",
-				Required:            true,
-			},
-			"allow_expired": schema.StringAttribute{
-				MarkdownDescription: "Allow expired user to be configured.",
-				Computed:            true,
-			},
-			"password": schema.StringAttribute{
-				MarkdownDescription: "User password.",
-				Computed:            true,
-			},
-			"password_encryption_type": schema.StringAttribute{
-				MarkdownDescription: "Password encryption type.",
-				Computed:            true,
-			},
-			"roles": schema.ListNestedAttribute{
-				MarkdownDescription: "User roles.",
+			"users": schema.ListNestedAttribute{
+				MarkdownDescription: "List of users.",
 				Computed:            true,
 				NestedObject: schema.NestedAttributeObject{
 					Attributes: map[string]schema.Attribute{
 						"name": schema.StringAttribute{
-							MarkdownDescription: "Role name.",
+							MarkdownDescription: "User name.",
 							Computed:            true,
+						},
+						"allow_expired": schema.StringAttribute{
+							MarkdownDescription: "Allow expired user to be configured.",
+							Computed:            true,
+						},
+						"password": schema.StringAttribute{
+							MarkdownDescription: "User password.",
+							Computed:            true,
+						},
+						"password_encryption_type": schema.StringAttribute{
+							MarkdownDescription: "Password encryption type.",
+							Computed:            true,
+						},
+						"roles": schema.ListNestedAttribute{
+							MarkdownDescription: "User roles.",
+							Computed:            true,
+							NestedObject: schema.NestedAttributeObject{
+								Attributes: map[string]schema.Attribute{
+									"name": schema.StringAttribute{
+										MarkdownDescription: "Role name.",
+										Computed:            true,
+									},
+								},
+							},
 						},
 					},
 				},
@@ -100,7 +108,7 @@ func (d *UserDataSource) Schema(ctx context.Context, req datasource.SchemaReques
 	}
 }
 
-func (d *UserDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
+func (d *UserManagementDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
 	}
@@ -111,8 +119,8 @@ func (d *UserDataSource) Configure(_ context.Context, req datasource.ConfigureRe
 // End of section. //template:end model
 
 // Section below is generated&owned by "gen/generator.go". //template:begin read
-func (d *UserDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-	var config User
+func (d *UserManagementDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	var config UserManagement
 
 	// Read config
 	diags := req.Config.Get(ctx, &config)
@@ -130,7 +138,7 @@ func (d *UserDataSource) Read(ctx context.Context, req datasource.ReadRequest, r
 	}
 
 	queries := []func(*nxos.Req){}
-	queries = append(queries, nxos.Query("rsp-subtree-depth", "2"))
+	queries = append(queries, nxos.Query("rsp-subtree-depth", "3"))
 	res, err := device.Client.GetDn(config.getDn(), queries...)
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))