+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- BREAKING CHANGE: Rename `nxos_user` resource and data source to `nxos_user_management`, now targeting the `sys/userext` class instead of `sys/userext/user-[name]`
 - Add `nxos_bridge_domains` bulk resource
 - Add `nxos_vrfs` bulk resource
 - Add `nxos_physical_interfaces` bulk resource